requirements.txt
python-telegram-bot[webhooks]>=20.0
requests>=2.31.0
aiohttp>=3.8.1
